requirements.txt
fastapi
uvicorn
sqlalchemy
pymongo
pydantic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
